feat: grid canvas limited height + merging button change
